spent all day trying to get authentication to work. Firebase is sendi…

+++ MadWrld/client/src/components/user/EditUser.tsx
@@ -3,19 +3,22 @@ import { useParams, useNavigate } from "react-router-dom";
 import { getUserById, deleteUser, editUserInfo, getAllUserTypes } from "../../modules/auth/authManager";
 import { Form, FormGroup, Input, Label } from "reactstrap"
 import IUser from "../../interfaces/IUser";
+import { useAuth } from "../../modules/auth/authContext";
         
-const EditUser: React.FC<{ userProfile: IUser | null}> = ({userProfile}) => {
+const EditUser: React.FC = () => {
     const navigate = useNavigate();
+    const { userProfile } = useAuth();
     const params = useParams<{ id: string }>();
-    const [ oldUser, setOldUser ] = useState<IUser>({id: 0, firstName: "", lastName: "", email: "", userTypeId: 0, firebaseUUID: ""});
-    const [ newUser, setNewUser ] = useState<IUser>(oldUser);
+    const [ user, setUser ] = useState<IUser>({ id: 0, firstName: '', lastName: '', email: '', userTypeId: 0});
     const [ userTypes, setUserTypes ] = useState<{id: number; name: string}[]>([]);
     const userId = params.id ? parseInt(params.id, 10) : 0;
+    const [oldUser, setOldUser] = useState<IUser | null>(null);
 
     useEffect(() => {
         const getUser = async () => {
             try {
                 const userData = await getUserById(userId);
+                setUser(userData);
                 setOldUser(userData);
             } catch (error) {console.error("Error fetching user data: ", error);}
         };
@@ -35,29 +38,31 @@ const EditUser: React.FC<{ userProfile: IUser | null}> = ({userProfile}) => {
         }
     }, [userId]);
 
-    useEffect(() => {
-        setNewUser(oldUser);
-    }, [oldUser]);
-
     const submitChanges: React.FormEventHandler<HTMLFormElement> = (e) => {
         e.preventDefault();
-          editUserInfo(newUser)
+          editUserInfo(user)
           .then(() => window.alert('Changes have been made!'))
-          .then(() => {navigate(`/users/${newUser.id}`)})
+          .then(() => {navigate(`/users/${user.id}`)})
           .catch(error => console.error("Error editing user info: ", error));
     };
     
     const handleDelete: React.MouseEventHandler<HTMLButtonElement> = (e) => {
         e.preventDefault();
-        const confirmation = window.confirm("Are you sure you want to delete this user?");
-        if (confirmation) {
-            deleteUser(oldUser.id)
-            .then(() => window.alert("User has been deleted!"))
-            .then(() => userProfile?.userTypeId === 1 ? navigate("/users") : navigate("/"))
-            .catch(error => console.error("Error deleting user: ", error));
+        if (oldUser && oldUser.id) {
+            const confirmation = window.confirm("Are you sure you want to delete this user?");
+            if (confirmation) {
+                deleteUser(oldUser.id)
+                .then(() => window.alert("User has been deleted!"))
+                .then(() => userProfile?.userTypeId === 1 ? navigate("/users") : navigate("/"))
+                .catch(error => console.error("Error deleting user: ", error));
+            }
         }
     };
 
+    if (!oldUser) {
+        return <div>loading...</div>
+    }
+    
     return (
     <>
         <section className="input-container">
@@ -70,18 +75,18 @@ const EditUser: React.FC<{ userProfile: IUser | null}> = ({userProfile}) => {
                 defaultValue={oldUser.firstName}
                 type="text"
                 onChange={(e) => {
-                    let copy = {...newUser}
+                    let copy = {...user}
                     copy.firstName = e.target.value 
-                    setNewUser(copy)}}
+                    setUser(copy)}}
                 />
                 <Label>Last Name</Label>
                 <Input
                 defaultValue={oldUser.lastName}
                 type="text"
                 onChange={(e) => {
-                    let copy = {...newUser}
+                    let copy = {...user}
                     copy.lastName = e.target.value 
-                    setNewUser(copy)}}
+                    setUser(copy)}}
                 />
             </FormGroup>
             <FormGroup>
@@ -90,9 +95,9 @@ const EditUser: React.FC<{ userProfile: IUser | null}> = ({userProfile}) => {
                 defaultValue={oldUser.email}
                 type="text"
                 onChange={(e) => {
-                    let copy = {...newUser}
+                    let copy = {...user}
                     copy.email = e.target.value 
-                    setNewUser(copy)}}
+                    setUser(copy)}}
                 />
             </FormGroup>
             <FormGroup>
@@ -103,9 +108,9 @@ const EditUser: React.FC<{ userProfile: IUser | null}> = ({userProfile}) => {
                     defaultValue={oldUser.userTypeId}
                     type="select"
                     onChange={(e) => {
-                        let copy = {...newUser}
+                        let copy = {...user}
                         copy.userTypeId = parseInt(e.target.value); 
-                        setNewUser(copy)}}
+                        setUser(copy)}}
                     >
                         {userTypes.map(ut => {
                         return (<option key={ut.id} value={ut.id}>{ut.name}</option>)
